#265: separated view model into their own package
