+++ app/src/androidTest/java/com/example/concordia_campus_guide/InstrumentalTests/ViewModelFactoryInstrumentalTest.java
@@ -1,11 +1,10 @@
 package com.example.concordia_campus_guide.InstrumentalTests;
 
 import com.example.concordia_campus_guide.Activities.MainActivity;
-import com.example.concordia_campus_guide.Activities.RoutesActivityViewModel;
-import com.example.concordia_campus_guide.Activities.SearchActivityViewModel;
-import com.example.concordia_campus_guide.Fragments.InfoCardFragment.InfoCardFragment;
-import com.example.concordia_campus_guide.Fragments.InfoCardFragment.InfoCardFragmentViewModel;
-import com.example.concordia_campus_guide.Fragments.LocationFragment.LocationFragmentViewModel;
+import com.example.concordia_campus_guide.ViewModels.RoutesActivityViewModel;
+import com.example.concordia_campus_guide.ViewModels.SearchActivityViewModel;
+import com.example.concordia_campus_guide.ViewModels.InfoCardFragmentViewModel;
+import com.example.concordia_campus_guide.ViewModels.LocationFragmentViewModel;
 import com.example.concordia_campus_guide.Helper.ViewModelFactory;
 
 import org.junit.Before;
