+++ app/src/main/java/com/example/concordia_campus_guide/ViewModels/SearchActivityViewModel.java
@@ -1,4 +1,4 @@
-package com.example.concordia_campus_guide.Activities;
+package com.example.concordia_campus_guide.ViewModels;
 
 import android.location.Location;
