+++ app/src/main/res/layout/point_of_interest_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".Fragments.POIFragment.POIFragment">
+    tools:context=".Fragments.POIFragment">
 
     <LinearLayout
         android:id="@+id/poiLinearLayout"
