+++ app/src/test/java/com/example/concordia_campus_guide/ViewModel/MainActivityViewModelTest.java
@@ -1,14 +1,12 @@
 package com.example.concordia_campus_guide.ViewModel;
 
-import com.example.concordia_campus_guide.Activities.MainActivityViewModel;
+import com.example.concordia_campus_guide.ViewModels.MainActivityViewModel;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.MockitoAnnotations;
 
 import static junit.framework.TestCase.assertEquals;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 public class MainActivityViewModelTest {
     private MainActivityViewModel mainActivityViewModel;
